+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Tables are now generated in the `app\Tables` directory.
   * Using this new way to manage table, you will have to display the view with `$table()` instead of `$table`.
   * Please note that even if is not documented anymore, the old behaviour is still functional in order to prevent any breaking change.
+* Removed the deprecated `->icon()` table column method.  
 
 ## [1.5.0](https://github.com/Okipa/laravel-table/compare/1.4.0...1.5.0)